refactor(cli): extract shared parseApprovalMode util
borrow, lend, and swap each carried a near-identical parseApprovalMode.
Pull it into utils/parseApprovalMode.js and import it everywhere. The
shared copy uses the dynamic APPROVAL_MODES.join(' or ') message; since
APPROVAL_MODES = ['exact', 'max'] it renders identically to swap's old
hardcoded text, so no observable change.
